requirements.txt
# BR10 Block Web - Dependências Python
# Versão: 3.0.0
# Data: 2026-02-08

# Framework Web
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Session==0.5.0
Flask-WTF==1.2.1
Werkzeug==3.0.1

# Banco de Dados
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Cache e Sessões
redis==5.0.1
hiredis==2.3.2

# Extração de PDF
PyPDF2==3.0.1
pdfplumber==0.10.3
pdfminer.six==20221105

# Validação e Segurança
email-validator==2.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
cryptography==42.0.0

# Utilitários
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0

# Monitoramento e Logs
psutil==5.9.7

# Desenvolvimento e Testes
pytest==7.4.4
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Produção
gunicorn==21.2.0
gevent==24.2.1
